% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{layColBind}
\alias{layColBind}
\title{Take two Layout objects and combine by rows.}
\usage{
layColBind(x, y, widths = c(1, 1), addmax = TRUE)
}
\arguments{
\item{x}{object of class Layout}

\item{y}{object of class Layout}
}
\description{
Take two Layout objects and combine by rows.
}
\examples{
l1 = layCreate(matrix(c(1:2),ncol = 2),widths=c(4,1))
l2 = layCreate(matrix(c(1:4),ncol = 2),widths=c(1,1))
lb = layColBind(l1,l2)
layoutShow(lb)
}
