% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{layRepCol}
\alias{layRepCol}
\title{Repeats columns in matrix.}
\usage{
layRepCol(x, fr)
}
\arguments{
\item{x}{matrix}

\item{fr}{numeric value or vector with number of repetition of given column.}
}
\description{
Repeats columns in matrix.
}
\examples{

layRepCol(matrix(1:4,ncol=2),c(4,2))
layRepCol(matrix(1:4,ncol=2),2)
}
