% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{layRepRow}
\alias{layRepRow}
\title{Repeats rows in matrix.}
\usage{
layRepRow(x, fr)
}
\arguments{
\item{x}{matrix}

\item{fr}{numeric value or vector with number of repetition of given row.}
}
\description{
Repeats rows in matrix.
}
\examples{

layRepRow(matrix(1:4,ncol=2),c(4,2))
layRepRow(matrix(1:4,ncol=2),2)
}
