% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{layRowBind}
\alias{layRowBind}
\title{Take two Layout objects and combine by rows.}
\usage{
layRowBind(x, y, heights = c(1, 1), addmax = TRUE)
}
\arguments{
\item{x}{object of class Layout}

\item{y}{object of class Layout}
}
\description{
Take two Layout objects and combine by rows.
}
\examples{
l1 = layCreate(matrix(c(1:2),ncol = 2),widths=c(4,1))
l2 = layCreate(matrix(c(1:4),ncol = 2),widths=c(1,1))
lb = layRowBind(l1,l2)
layoutShow(lb)

}
