% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{laySet}
\alias{laySet}
\title{Set custom layout.}
\usage{
laySet(layout)
}
\arguments{
\item{layout}{object of class Layout.}
}
\description{
Set custom layout.
}
\examples{

lplots = layCreate(matrix(1:2))
lpie   = layCreate(1)
lay = layColBind(lplots,lpie)
laySet(lay)
plot(1:10)
plot(1:10)
plot(1:20)
}
