% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitField.R
\name{laySplitField}
\alias{laySplitField}
\title{Split a selected field from layout using a schema from another layout.}
\usage{
laySplitField(lay, newlay, field)
}
\arguments{
\item{lay}{a Layout object.}

\item{newlay}{a Layout object used to split a field from \code{lay}.}

\item{field}{id of a field from \code{lay}.}
}
\description{
Split a selected field from layout using a schema from another layout.
}
\examples{

l1 <- layCreate(matrix(c(1:4), ncol = 2), widths = c(4, 1))
l2 <- layCreate(matrix(c(1:4), ncol = 2), widths = c(1, 1))
l3 <- laySplitField(l1, l2, 2)
layShow(l3)

}
