% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{phl_calc_fontsize}
\alias{phl_calc_fontsize}
\title{Calculate optimal fontsize and height of the cell for given height for flextable.}
\usage{
phl_calc_fontsize(data, height)
}
\arguments{
\item{data}{data.frame.}

\item{height}{single numeric value with desired height.}
}
\value{
A named numeric vector containing two elements:
\itemize{
  \item \code{fs} font size
  \item \code{height} of the single cell.
}
}
\description{
Calculate optimal fontsize and height of the cell for given height for flextable.
}
\examples{

x <- tail(iris, 10)[,c(1,5)]
phl_calc_fontsize(x, 5)

}
