% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{phl_with_gg}
\alias{phl_with_gg}
\title{add ggplot into layout placeholder}
\usage{
phl_with_gg(x, olay, id, value, ...)
}
\arguments{
\item{x}{rpptx object}

\item{olay}{an OfficerLayout object created using \code{\link{phl_layout}}}

\item{id}{an single integer with an id of the placeholder from \code{olay} object.}

\item{value}{a ggplot object}

\item{...}{other arguments passed to \code{\link{ph_with_gg_at}}}
}
\description{
add ggplot into layout placeholder
}
