% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{phl_with_table}
\alias{phl_with_table}
\title{add table into layout placeholder}
\usage{
phl_with_table(x, olay, id, value, ...)
}
\arguments{
\item{x}{rpptx object}

\item{olay}{an OfficerLayout object created using \code{\link{phl_layout}}}

\item{id}{an single integer with an id of the placeholder from \code{olay} object.}

\item{value}{a data.frame}

\item{...}{other arguments passed to \code{\link{ph_with_table_at}}}
}
\description{
add table into layout placeholder
}
