% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{print.CustomLayout}
\alias{print.CustomLayout}
\title{Print a CustomLayout object.}
\usage{
\method{print}{CustomLayout}(x, ...)
}
\arguments{
\item{x}{object of class CustomLayout.}

\item{...}{optional arguments to print or plot methods. Not used here.}
}
\description{
Print a CustomLayout object.
}
\examples{

lay  <- lay_new(matrix(1:4,nc=2),widths=c(3,2),heights=c(2,1))
lay2 <- lay_new(matrix(1:3))
cl <- lay_bind_col(lay,lay2, widths=c(3,1))
print(cl)

}
\seealso{
lay_new lay_show
}
