% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{print.OfficerCustomLayout}
\alias{print.OfficerCustomLayout}
\title{Print a OfficerCustomLayout object.}
\usage{
\method{print}{OfficerCustomLayout}(x, ...)
}
\arguments{
\item{x}{object of class OfficerCustomLayout}

\item{...}{optional arguments to print or plot methods. Not used here.}
}
\description{
Print a OfficerCustomLayout object.
}
\examples{

lay  <- lay_new(matrix(1:4,nc = 2),widths = c(3, 2),heights = c(2, 1))
lay2 <- lay_new(matrix(1:3))
cl <- lay_bind_col(lay,lay2, widths=c(3,1))
ofl <- phl_layout(cl, innerMargins = rep(0.1,4))
print(ofl)

}
\seealso{
lay_new lay_show phl_layout
}
