% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{lay_bind_row}
\alias{lay_bind_row}
\alias{layRowBind}
\title{Take two Layout objects and combine by rows.}
\usage{
lay_bind_row(x, y, heights = c(1, 1), addmax = TRUE)

layRowBind(x, y, heights = c(1, 1), addmax = TRUE)
}
\arguments{
\item{x}{object of class Layout}

\item{y}{object of class Layout}

\item{heights}{a vector with relative heights used in combining 
the x and y layouts.}

\item{addmax}{if true (default) the ids of the plots in the second 
layout will be shifted by the number of plots in 
the first layout.}
}
\description{
Take two Layout objects and combine by rows.
}
\examples{
l1 = lay_new(matrix(c(1:2),ncol = 2),widths=c(4,1))
l2 = lay_new(matrix(c(1:4),ncol = 2),widths=c(1,1))
lb = lay_bind_row(l1,l2)
lay_show(lb)

}
