% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{phl_with_plot}
\alias{phl_with_plot}
\title{add plot into layout placeholder}
\usage{
phl_with_plot(x, olay, id, plotFnc, ...)
}
\arguments{
\item{x}{rpptx object}

\item{olay}{an OfficerLayout object created using \code{\link{phl_layout}}}

\item{id}{an single integer with an id of the placeholder from \code{olay} object.}

\item{plotFnc}{a function which creates a plot when called.}

\item{...}{other arguments passed to \code{\link{png}} function.}
}
\description{
add plot into layout placeholder
}
