% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{phl_with_vg}
\alias{phl_with_vg}
\title{add a plot as vector graphics into layout placeholder}
\usage{
phl_with_vg(x, olay, id, code, ggobj = NULL, ...)
}
\arguments{
\item{x}{rpptx object}

\item{olay}{an OfficerLayout object created using \code{\link{phl_layout}}}

\item{id}{an single integer with an id of the placeholder from \code{olay} object.}

\item{code}{plot instructions.}

\item{ggobj}{ggplot objet to print. Argument code will be ignored if this argument is supplied.}

\item{...}{other arguments passed to \code{\link{dml_pptx}}}
}
\description{
add a plot as vector graphics into layout placeholder
}
